Got the sentiment analysis done: updated the datastore and js

+++ src/main/java/com/google/codeu/data/Datastore.java
@@ -69,7 +69,7 @@ public class Datastore {
         UUID id = UUID.fromString(idString);
         String text = (String) entity.getProperty("text");
         long timestamp = (long) entity.getProperty("timestamp");
-        float score = (float) entity.getProperty("score");
+        double score = (double) entity.getProperty("score");
 
         Message message = new Message(id, user, text, timestamp, score);
         messages.add(message);
@@ -102,7 +102,7 @@ public class Datastore {
         String user = (String) entity.getProperty("user");
         String text = (String) entity.getProperty("text");
         long timestamp = (long) entity.getProperty("timestamp");
-        float score = (float) entity.getProperty("score");
+        double score = (double) entity.getProperty("score");
 
         Message message = new Message(id, user, text, timestamp, score);
         messages.add(message);
